cd: fix deployment issue with dependencies
Work around build/deploy issues by removing the use of Firebase Admin by using
the Firestore REST API instead. This allows changing to the "serverless" build
target instead of "experimental-serverless-trace" which lets the app be
deployable while keeping electron as a dependency.

The downside is not being able to verify user tokens server side since
Firebase's Admin SDK is required for that. This functionality isn't currently
need but would be good to have in the future.

Ideally electron should not be a dependency. However, for now, it is needed by
@mapbox/mapbox-gl-geocoder for geocoding. Perhaps we can use
"experimental-serverless-trace" (or the useServerlessTraceTarget input for
serverless.yml config) eventually, after directly fetching from Mapbox's
geocoding API which would remove the need of @mapbox/mapbox-gl-geocoder and
electron.

- firebase-admin works with "experimental-serverless-trace", not "serverless"
  build target
- @mapbox/mapbox-gl-geocoder relies on @mapbox/mapbox-sdk, which relies on
  electron
- electron generally shouldn't be needed and is very large as a dependency
- electron breaks the deploy to Lambda functions due to their size limit when
  using "experimental-serverless-trace" target but seems fine with just the
  "serverless" target

See:
- https://github.com/Wykks/ngx-mapbox-gl/issues/219
- https://github.com/netlify/next-on-netlify/issues/68#issuecomment-719994235